showroomapi/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
astroid==2.8.6
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
black==23.1a1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cssselect2==0.7.0
daphne==4.0.0
dill==0.3.6
Django==4.1.4
django-auto-prefetching==0.2.12
django-cors-headers==3.13.0
django-doctor==2.6.0
django-filter==22.1
django-rest-swagger==2.2.0
django-watchman==1.3.0
django_debug_toolbar==3.8.1
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dparse==0.6.2
drf-ujson==1.2
drf-ujson2==1.7.2
drf-yasg2==1.19.4
exceptiongroup==1.1.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.11.2
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.8.0
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
mdurl==0.1.2
msgpack==1.0.4
mypy-extensions==0.4.3
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==21.3
pathspec==0.11.0
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
psutil==5.9.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.14.0
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
pylint==2.11.1
pyngrok==5.2.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPDF3==1.0.6
pytest==7.2.0
pytest-django==4.5.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywatchman==1.4.1
PyYAML==6.0
qrcode==7.3.1
redis==4.4.0
reportlab==3.6.12
requests==2.28.1
requests-auth==6.0.0
rich==13.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.1
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.13
urllib3-secure-extra==0.1.0
webencodings==0.5.1
whitenoise==6.3.0
wrapt==1.13.3
xhtml2pdf==0.2.8
zope.interface==5.5.2
